requirements.txt
appdirs==1.4.4
asgiref
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cssselect==1.1.0
cssutils==2.2.0
distlib==0.3.1
Django==3.0.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djongo==1.3.4
idna==2.10
libLAS==1.8.1
lxml==4.6.2
premailer==3.7.0
PyJWT==1.7.1
pytz
requests==2.25.1
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.3
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
wincertstore==0.2
yagmail==0.14.245
